Stop one system's brand rotation from preventing another
Twitter's brand rotation should be independent of the success of
Gravatar's. To do this, we tell them to continue on error, but that
would hide any errors, making it seem like the workflow succeeded when
it didn't completely. So we add a step to check the status of each
rotation and fail the workflow if either step failed.